[fix] Move FileDownloadRequest to /abstractions/fileDownload
