+++ libs/common/src/abstractions/fileDownload/fileDownload.service.ts
@@ -1,4 +1,4 @@
-import { FileDownloadRequest } from "../models/domain/fileDownloadRequest";
+import { FileDownloadRequest } from "./fileDownloadRequest";
 
 export abstract class FileDownloadService {
   download: (request: FileDownloadRequest) => void;
